requirements.txt
elasticdeform==0.5.0
keras==2.15.0
matplotlib==3.7.1
neurite==0.2
nibabel==5.2.1
numpy==1.23.5
open3d==0.18.0
opencv-python==4.7.0.72
pandas==2.0.1
Pillow==9.5.0
scikit-image==0.22.0
SimpleITK==2.3.1
tensorflow==2.15.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0